Moved command controllers in separate package
