+++ pkg/controllers/command/daemon.go
@@ -14,17 +14,27 @@
  *    limitations under the License.
  */
 
-package controllers
+package command
 
 import (
-	"fmt"
+	"log/slog"
+	"net"
+	"strconv"
+	"time"
+
+	"github.com/corelayer/netscaleradc-acme-go/pkg/models/config"
 )
 
-type ConfigureCertificateCommand struct {
+type Daemon struct {
+	Config config.Daemon
 }
 
-func (c ConfigureCertificateCommand) Execute() error {
-	var err error
-	fmt.Println("Configure certificate settings")
-	return err
+func (c Daemon) Execute() error {
+	if _, err := net.Listen("tcp", c.Config.Address+":"+strconv.Itoa(c.Config.Port)); err != nil {
+		slog.Error("a daemon is already running on the same address")
+		return err
+	}
+	slog.Info("Running daemon")
+	time.Sleep(10 * time.Second)
+	return nil
 }
